refactor: rename troms identifier from tfk to troms
