+++ packages/assets/src/generate.ts
@@ -18,7 +18,7 @@ export const vaildOrgIds = [
   ThemeVariant.AtB,
   ThemeVariant.Nfk,
   ThemeVariant.FRAM,
-  ThemeVariant.TFK,
+  ThemeVariant.TROMS,
 ];
 export const searchGlob = '**/*.{svg,png,jpg,jpeg,ico}';
 export const searchGlobSvg = '**/*.svg';
