+++ packages/theme/src/themes/index.ts
@@ -1,4 +1,4 @@
 export {default as AtBThemes} from './atb-theme/theme';
 export {default as NfkThemes} from './nfk-theme/theme';
 export {default as FRAMThemes} from './fram-theme/theme';
-export {default as TFKThemes} from './tfk-theme/theme';
+export {default as TromsThemes} from './troms-theme/theme';
